# api-description/ofrep.openapi.yaml
...
openapi: 3.1.0
servers:
  - url: /
info:
  version: 0.3.0
  title: Bucketeer OpenFeature Remote Evaluation Protocol (OFREP)
  description: |
    Bucketeer's OFREP 0.3 evaluation API, derived from the official OpenFeature
    protocol schema. It evaluates flags with Bucketeer's existing evaluator and
    requires an enabled SDK_SERVER API key.

    Single evaluation makes a bounded, best-effort attempt to record an exposure
    event. Bulk evaluation does not record exposures or advertise event streams.
    Flag management, goal tracking, CORS, retries, and rate limiting are outside
    this API.

    See https://openfeature.dev/docs/reference/other-technologies/ofrep/ and
    ../docs/mechanism/ofrep.md for usage and implementation details.

  contact:
    url: https://github.com/open-feature/protocol
  license:
    identifier: Apache-2.0
    name: Apache 2.0
security:
  - ApiKeyAuth: []
  - BearerAuth: []
  - BucketeerAuthorizationAuth: []
...
paths:
  /ofrep/v1/evaluate/flags/{key}:
    post:
      tags: [OFREP Core]
      summary: Evaluate A Single Feature Flag
      description: |
        Evaluates a single feature flag by its key. This endpoint is used by **server-side providers**
        for dynamic context evaluation, where each evaluation request includes the evaluation context.

        The endpoint returns the evaluated flag value along with metadata including the evaluation
        reason, variant, and any flag-specific metadata. The flag value can be one of several types:
        boolean, string, integer, float, or object.

        **Use Case**: Server-side applications where evaluation context may change between requests
        and real-time targeting decisions are required.
      operationId: evaluateFlag
      parameters:
        - name: key
          in: path
          required: true
          description: The unique identifier (key) of the feature flag
          schema:
            type: string
          example: discount-banner
      requestBody:
        required: true
        description: Evaluation request containing the context for flag evaluation
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/evaluationRequest"
            example:
              context:
                targetingKey: user-123
                email: user@example.com
                custom-plan: premium
                country: CA
      responses:
        "200":
          description: Successful flag evaluation. Returns the evaluated flag value with metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/serverEvaluationSuccess"
              example:
                key: discount-banner
                value: true
                reason: TARGETING_MATCH
                variant: enabled
        "400":
          description: Bad evaluation request. The request is malformed or contains invalid context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/evaluationFailure"
              example:
                key: my-flag
                errorCode: INVALID_CONTEXT
                errorDetails: "Context is missing required targetingKey property"
        "404":
          description: Flag not found. The specified flag key does not exist in the flag management system.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/flagNotFound"
              example:
                key: non-existent-flag
                errorCode: FLAG_NOT_FOUND
                errorDetails: "Flag 'non-existent-flag' was not found"
        "401":
          description: Unauthorized. Authentication credentials are missing, invalid, or expired.
        "403":
          description: Forbidden. The client does not have permission to access the requested resource.
        "429":
          description: Too Many Requests. Rate limit has been exceeded.
          headers:
            Retry-After:
              description: |
                Indicates when to retry the request again. Can be either a date-time string
                or a number of seconds to wait.
              schema:
                type: string
                format: date-time
                example: "2024-02-07T12:00:00Z"
        "500":
          description: Internal Server Error. An unexpected error occurred on the server that prevented flag evaluation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalErrorResponse"
              example:
                errorDetails: "An internal server error occurred while processing the request"
  /ofrep/v1/evaluate/flags:
    post:
      tags: [OFREP Core]
      summary: Bulk Evaluate All Feature Flags
      description: |
        Evaluates all feature flags in a single request using a static context. This endpoint is
        used by **client-side providers** for static context evaluation, where all flags are evaluated
        once and then cached locally for subsequent use.

        The endpoint returns an array of all flag evaluations, where each flag can be either a
        successful evaluation or an evaluation failure. The response includes an ETag header for
        cache validation, allowing clients to use the `If-None-Match` header to avoid downloading
        an unchanged response body. Bucketeer still evaluates the flags before comparing the ETag.
      operationId: evaluateFlagsBulk
      parameters:
        - in: header
          name: If-None-Match
          description: |
            Optional ETag value from a previous bulk evaluation response. If provided and the
            ETag matches the current flag set, the server will return a 304 Not Modified response,
            indicating that flags haven't changed since the last evaluation.
          schema:
            type: string
          required: false
          example: abc123xyz
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/bulkEvaluationRequest"
            example:
              context:
                targetingKey: user-456
                email: user@example.com
                plan: free
                country: CA
      responses:
        "200":
          description: Successful bulk evaluation.
          headers:
            ETag:
              schema:
                type: string
              description: |
                Entity tag (ETag) representing the current state of all flags. Clients should
                include this value in subsequent requests using the `If-None-Match` header for
                cache validation.
              example: abc123xyz
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bulkEvaluationSuccess"
              example:
                flags:
                  - key: discount-banner
                    value: true
                    reason: TARGETING_MATCH
                    variant: enabled
                  - key: theme-color
                    value: "blue"
                    reason: STATIC
                    variant: default
                  - key: non-existent-flag
                    errorCode: FLAG_NOT_FOUND
                    errorDetails: "Flag 'non-existent-flag' was not found"
        "304":
          description: |
            Not Modified. The flags haven't changed since the last evaluation (ETag matches). No response body is returned.
        "400":
          description: Bad evaluation request. The request is malformed or contains invalid context.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/bulkEvaluationFailure"
              example:
                errorCode: INVALID_CONTEXT
                errorDetails: "Context is missing required targetingKey property"
        "401":
          description: Unauthorized. Authentication credentials are missing, invalid, or expired.
        "403":
          description: Forbidden. The client does not have permission to access the requested resource.
        "429":
          description: Too Many Requests. Rate limit has been exceeded.
          headers:
            Retry-After:
              description: |
                Indicates when to retry the request again. Can be either a date-time string
                or a number of seconds to wait.
              schema:
                oneOf:
                  - type: string
                    format: date-time
                  - type: integer
                    description: Number of seconds to wait before retrying
              example: "2024-02-07T12:00:00Z"
        "500":
          description: Internal server error. An unexpected error occurred on the server that prevented flag evaluation.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/generalErrorResponse"
              example:
                errorDetails: "An internal server error occurred while processing the request"
components:
  securitySchemes:
    BearerAuth:
      description: |
        An enabled Bucketeer SDK_SERVER API key in the format
        `Authorization: Bearer <key>`.
      type: http
      scheme: bearer
    ApiKeyAuth:
      description: |
        An enabled Bucketeer SDK_SERVER API key in the `X-API-Key` header.
      type: apiKey
      in: header
      name: X-API-Key
    BucketeerAuthorizationAuth:
      description: |
        Backwards-compatible Bucketeer authentication using a raw SDK_SERVER API
        key in the `Authorization` header, without the Bearer scheme.
      type: apiKey
      in: header
      name: Authorization
  schemas:
...
    evaluationRequest:
      description: |
        Request body for single feature flag evaluation. Contains the evaluation context that will be
        used to evaluate the feature flag.
      type: object
      required:
        - context
      properties:
        context:
          $ref: "#/components/schemas/context"
    serverEvaluationSuccess:
      allOf:
        - $ref: "#/components/schemas/evaluationSuccess"
...
    evaluationFailure:
      description: |
        Feature flag evaluation failure response. Returned when a feature flag evaluation fails, such as when
        the context is invalid, required targeting keys are missing, or parsing errors occur.
      type: object
      properties:
        key:
          $ref: "#/components/schemas/key"
        errorCode:
          type: string
          description: |
            OpenFeature compatible error code indicating the type of error. See
            https://openfeature.dev/specification/types#error-code for error code definitions.
          enum: [PARSE_ERROR, TARGETING_KEY_MISSING, INVALID_CONTEXT, GENERAL]
        errorDetails:
          $ref: "#/components/schemas/errorDetails"
        metadata:
          allOf:
            - $ref: "#/components/schemas/metadata"
            - $ref: "#/components/schemas/flagMetadataDescription"
            - $ref: "#/components/schemas/flagMetadataExamples"
      required:
        - key
        - errorCode
    flagNotFound:
      description: |
        Feature flag not found response. Returned when a requested feature flag key does not exist in the
        flag management system. This can be an expected response in certain flag management systems.
        This is distinct from evaluation failures, which occur when a flag exists but cannot be evaluated.
      type: object
      properties:
        key:
          $ref: "#/components/schemas/key"
        errorCode:
          type: string
          description: Error code indicating the flag was not found
          enum: [FLAG_NOT_FOUND]
        errorDetails:
          $ref: "#/components/schemas/errorDetails"
        metadata:
          allOf:
            - $ref: "#/components/schemas/metadata"
            - $ref: "#/components/schemas/flagMetadataDescription"
            - $ref: "#/components/schemas/flagMetadataExamples"
      required:
        - key
        - errorCode
    generalErrorResponse:
      description: |
        General error response for unexpected server errors (500 status). Used when the
        server encounters an internal error that prevents flag evaluation.
      type: object
      properties:
        errorDetails:
          $ref: "#/components/schemas/errorDetails"
    key:
      type: string
      description: |
        The unique identifier (key) of the feature flag. This is used to reference a specific
        flag in the flag management system.
      example: discount-banner
...